TypeScript port, and the integer-division bug it was hiding
The port generates byte-identical SQL: 51 parity fixtures plus the golden
snapshot, compared as text and as parameter arrays, with 197 tests of its own.

Running it against a real server is what caught the bug. Every score came back
0.000000 and the ranking collapsed to insertion order, while the SQL was
character-for-character the same as Python's. The difference was the parameters:
JavaScript has one number type, so the driver sends 1 where Python sends 1.0,
Postgres infers integer for the weight and for k, and 1 / (60 + rank) becomes
integer division. Nothing errors. The query succeeds, returns the right rows,
and scores every one of them zero.

The parity checker could not see it, because 1 and 1.0 compare equal.

Fixed in the SQL rather than in either client, by casting the arithmetic
parameters to float8 so the expression no longer depends on a driver's type
inference at all. A regression test binds every numeric parameter as an integer
-- exactly what the TypeScript package does on the wire -- and asserts the
scores are non-zero.

+++ python/tests/golden/canonical_search.sql
@@ -23,10 +23,10 @@ scored AS (
     SELECT coalesce(v.id, t.id) AS id,
            v.rank AS vector_rank,
            v.distance AS vector_distance,
-           coalesce($11 / ($12 + v.rank), 0) AS vector_contribution,
+           coalesce($11::float8 / ($12::float8 + v.rank), 0) AS vector_contribution,
            t.rank AS text_rank,
            t.score AS text_score,
-           coalesce($13 / ($12 + t.rank), 0) AS text_contribution
+           coalesce($13::float8 / ($12::float8 + t.rank), 0) AS text_contribution
     FROM vector_candidates v
     FULL OUTER JOIN text_candidates t ON v.id = t.id
 ),
